requirements.txt
google-adk>=0.1.0
streamlit
pandas
plotly
python-dotenv
nest_asyncio
litellm
# V2 Dependencies
pdfplumber>=0.10.0
python-docx>=1.0.0
audio-recorder-streamlit>=0.0.8
edge-tts>=6.1.0
groq>=0.4.0
